+++ CHANGELOG.rst
@@ -32,6 +32,8 @@ Added
 - add a feature for each regex in the training set for crf_entity_extractor
 - Current training processes count for server and projects.
 - the ``/version`` endpoint returns a new field ``minimum_compatible_version``
+- added logging of intent prediction errors to evaluation script
+- added histogram of confidence scores to evaluation script
 
 Changed
 -------
@@ -40,6 +42,7 @@ Changed
 - Allow multiple training processes per project
 - Changed AlreadyTrainingError to MaxTrainingError. The first one was used to indicate that the project was already training. The latest will show an error when the server isn't able to training more models.
 - ``Interpreter.ensure_model_compatibility`` takes a new parameters for the version to compare the model version against
+- confusion matrix plot gets saved to file automatically during evaluation
 
 Removed
 -------